itersolve: Convert iterative solver to use "secant method"
The previous code calculates each step time via an "exponential
search" followed by a "false position with Illinois algorithm" search.
Replace with a "secant method" with "bounds checks" search.  This
simplifies the code, improves the performance, and does a better job
of finding steps immediately prior to a direction change.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>